zaakafhandelcomponent-1208 - Updaten CMMN model
