+++ src/main/java/net/atos/zac/app/admin/model/RESTHumanTaskParameters.java
@@ -16,8 +16,4 @@ public class RESTHumanTaskParameters {
     public Integer doorlooptijd;
 
     public RESTGroup defaultGroep;
-
-    public RESTHumanTaskParameters() {
-    }
-
 }
